refactor(tailwind): extract generic icon pack factory
Pulls the lucide-specific tailwind plugin into a reusable
`iconPackPlugin({ prefix, iconsDir, normalizeStrokeWidth, defaultColor })`
factory. The lucide plugin is now a thin wrapper that preserves the
existing default export, so downstream tailwind configs are unaffected.

Opens the door to additional icon packs (tabler, heroicons, custom in-repo
packs) by registering them with their own class prefix.
